files w/o jsx have a .js extension now. Refactored reducers. Added linting capability but there were 3000+ errors so we'll so how that goes.
