+++ .babelrc
@@ -1,1 +1,1 @@
-{ "presets": ["react", "es2015"] }
+{ "presets": ["react", "stage-0", "es2015"] }
