+++ client/reducers/authentication.js
@@ -8,7 +8,7 @@ import { UPDATE_AUTHENTICATION } from '../actionTypes'
  * @param {string} action
  * @return new value for this store property or default state
  */
-export default function(state = initialState, action) {
+export default function(state = initialState.authenticated, action) {
 
   switch(action.type){
     case UPDATE_AUTHENTICATION:
