+++ client/reducers/lastChanged.js
@@ -1,17 +1,18 @@
 import initialState from '../store/initialState'
-import { UPDATE_TABLE_DATA } from '../actionTypes'
+import { UPDATE_LAST_CHANGED } from '../actionTypes'
 
 /**
- * Reducer for table data
- * @function tableReducer
+ * Reducer for last changed property in the store
+ * @function lastChangedReducer
  * @param {Object} state
  * @param {string} action
  * @return new value for this store property or default state
  */
-export default function(state = initialState, action) {
+
+export default function(state = initialState.lastChanged, action) {
 
   switch(action.type){
-    case UPDATE_TABLE_DATA:
+    case UPDATE_LAST_CHANGED:
       return action.data
       break
