+++ client/reducers/tab.js
@@ -8,7 +8,7 @@ import { CHANGE_TAB } from '../actionTypes'
  * @param {string} action
  * @return new value for this store property or default state
  */
-export default function(state = initialState, action) {
+export default function(state = initialState.tab, action) {
 
   switch(action.type){
     case CHANGE_TAB:
