+++ client/store/initialState.js
@@ -11,4 +11,5 @@ export default {
   detail: {},
   notifications: [],
   tab: 0,
+  userSettings: {}
 }
