+++ package.json
@@ -4,6 +4,7 @@
   "description": "Invenstory provides Amazon sellers with a useful information dashboard.",
   "main": "server.js",
   "scripts": {
+    "lint": "eslint ./",
     "postinstall": "mkdir -p dist && cp favicon.ico dist/; cp -R assets dist/",
     "pretest": "./node_modules/.bin/knex seed:run",
     "test": "./node_modules/.bin/mocha --compilers .:client/test/compiler.js,.:babel-core/register server/test client/test/ --recursive",
@@ -62,14 +63,21 @@
     "apidoc": "^0.15.1",
     "babel-cli": "^6.7.5",
     "babel-core": "^6.7.6",
+    "babel-eslint": "^6.1.2",
     "babel-loader": "^6.2.4",
     "babel-polyfill": "^6.7.4",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
+    "babel-preset-stage-0": "^6.5.0",
     "babel-watch": "^2.0.2",
     "chai": "^3.5.0",
     "chai-enzyme": "^0.4.2",
     "css-loader": "^0.23.1",
+    "eslint": "^3.1.1",
+    "eslint-config-airbnb": "^9.0.1",
+    "eslint-plugin-import": "^1.12.0",
+    "eslint-plugin-jsx-a11y": "^2.0.1",
+    "eslint-plugin-react": "^5.2.2",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
     "html-webpack-plugin": "^2.15.0",
